abstract away
