+++ architecture/src/main/java/com/xfinity/blueprint/architecture/Architect.kt
@@ -4,6 +4,7 @@ import `in`.srain.cube.views.ptr.PtrClassicFrameLayout
 import android.view.View
 import androidx.appcompat.app.ActionBar
 import com.xfinity.blueprint.ComponentRegistry
+import com.xfinity.blueprint.ViewAdapter
 import com.xfinity.blueprint.presenter.ScreenPresenter
 import com.xfinity.blueprint.view.ScreenViewDelegate
 
@@ -28,7 +29,7 @@ abstract class DefaultArchitect<out T : DefaultScreenView>(override val componen
         ptrFrame = layout.findViewById(R.id.ptr_frame)
         supportActionBar = actionBar
 
-        screenViewDelegate = ScreenViewDelegate(componentRegistry, loadingDots, recyclerView)
+        screenViewDelegate = ScreenViewDelegate(componentRegistry, ViewAdapter(recyclerView, loadingDots))
 
         val screenView = getScreenView()
         presenter.attachView(screenView)
