+++ library/src/main/java/com/xfinity/blueprint/view/ScreenViewDelegate.kt
@@ -11,17 +11,14 @@
 
 package com.xfinity.blueprint.view
 
-import android.view.View
 import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.SCROLL_STATE_IDLE
 import com.xfinity.blueprint.ComponentAdapter
 import com.xfinity.blueprint.ComponentRegistry
+import com.xfinity.blueprint.ViewRegistry
 import com.xfinity.blueprint.model.Component
 
 open class ScreenViewDelegate(componentRegistry: ComponentRegistry,
-                              private var loadingView: View? = null,
-                              private val recyclerView: RecyclerView) : ScreenView {
+                              private val viewRegistry: ViewRegistry) : ScreenView {
     val componentAdapter = ComponentAdapter(componentRegistry)
 
     override fun refresh() {
@@ -79,11 +76,11 @@ open class ScreenViewDelegate(componentRegistry: ComponentRegistry,
     }
 
     override fun showLoading() {
-        loadingView?.visibility = View.VISIBLE
+        viewRegistry.showLoading()
     }
 
     override fun hideLoading() {
-        loadingView?.visibility = View.GONE
+        viewRegistry.hideLoading()
     }
 
     /**
@@ -95,14 +92,9 @@ open class ScreenViewDelegate(componentRegistry: ComponentRegistry,
      * @param newComponents A List of the components the screen should be updated to display
      */
     override fun updateComponents(newComponents: List<Component>) {
-        val behavior = {
+        viewRegistry.updateComponents(newComponents) {
             updateComponents(newComponents) { oldComponents -> DefaultComponentDiffCallback(oldComponents, newComponents) }
         }
-        if (!recyclerView.isComputingLayout) {
-            behavior.invoke()
-            return
-        }
-        recyclerView.addOnScrollListener(RecyclerViewScrollHandler(behavior))
     }
 
     /**
@@ -123,15 +115,6 @@ open class ScreenViewDelegate(componentRegistry: ComponentRegistry,
         componentAdapter.clear(false)
         componentAdapter.addComponents(newComponents, false)
         diff.dispatchUpdatesTo(componentAdapter)
-    }
-
-    private inner class RecyclerViewScrollHandler(private val behavior: () -> Unit) : RecyclerView.OnScrollListener() {
-
-        override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
-            if (newState == SCROLL_STATE_IDLE && !recyclerView.isComputingLayout) {
-                behavior.invoke()
-            }
-        }
     }
 }
